requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
numpy==1.26.4
pandas==2.1.3
openpyxl==3.1.2
Pillow==10.1.0
gunicorn